requirements.txt
aniso8601==9.0.1
APScheduler==3.10.1
bidict==0.22.1
bson==0.5.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
docx2pdf==0.1.8
Flask==2.2.3
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.3
fpdf==1.7.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
Pillow==9.5.0
pycparser==2.21
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
requests==2.28.2
six==1.16.0
tqdm==4.65.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.14.0
zope.event==4.6
zope.interface==6.0
